Unify Command Bench tests into one class with methods
